Make TransitCurrencyTest run as unit test
