+++ src/androidTest/kotlin/au/id/micolous/metrodroid/test/TransitCurrencyTest.kt
@@ -18,12 +18,14 @@
  */
 package au.id.micolous.metrodroid.test
 
+import android.os.Build
 import android.text.Spanned
-
-import org.hamcrest.Matchers
-import org.junit.Test
+import android.text.style.TtsSpan
 
 import au.id.micolous.metrodroid.transit.TransitCurrency
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
 /**
  * Tests the currency formatter.
@@ -61,8 +63,7 @@ class TransitCurrencyTest : BaseInstrumentedTest() {
 
         // May be "USD12.34", "U$12.34" or "US$12.34".
         val usd = TransitCurrency.USD(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(usd, Matchers.startsWith("U"))
-        assertSpannedThat(usd, Matchers.endsWith("12.34"))
+        checkSpanned(usd, listOf("U$12.34", "US$12.34", "USD12.34"))
         assertTtsMarkers("USD", "12", "34", usd)
     }
 
@@ -77,13 +78,12 @@ class TransitCurrencyTest : BaseInstrumentedTest() {
 
         // May be "$12.34", "U$12.34" or "US$12.34".
         val usd = TransitCurrency.USD(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(usd, Matchers.endsWith("$12.34"))
+        checkSpanned(usd, listOf("$12.34", "U$12.34", "US$12.34"), "USD12.34")
         assertTtsMarkers("USD", "12", "34", usd)
 
         // May be "A$12.34" or "AU$12.34".
         val aud = TransitCurrency.AUD(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(aud, Matchers.startsWith("A"))
-        assertSpannedThat(aud, Matchers.endsWith("$12.34"))
+        checkSpanned(aud, listOf("A$12.34", "AU$12.34"), "AUD12.34")
         assertTtsMarkers("AUD", "12", "34", aud)
 
         val gbp = TransitCurrency(1234, "GBP").formatCurrencyString(true).spanned
@@ -92,7 +92,7 @@ class TransitCurrencyTest : BaseInstrumentedTest() {
 
         // May be "¥1,234" or "JP¥1,234".
         val jpy = TransitCurrency.JPY(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(jpy, Matchers.endsWith("¥1,234"))
+        checkSpanned(jpy, listOf("¥1,234", "JP¥1,234"), "JPY1,234")
         assertTtsMarkers("JPY", "1234", null, jpy)
     }
 
@@ -110,17 +110,16 @@ class TransitCurrencyTest : BaseInstrumentedTest() {
 
         // May be "A$12.34" or "AU$12.34".
         val aud = TransitCurrency.AUD(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(aud, Matchers.startsWith("A"))
-        assertSpannedThat(aud, Matchers.endsWith("$12.34"))
+        checkSpanned(aud, listOf("A$12.34", "AU$12.34"), "AUD12.34")
         assertTtsMarkers("AUD", "12", "34", aud)
 
         val gbp = TransitCurrency(1234, "GBP").formatCurrencyString(true).spanned
-        assertSpannedEquals("£12.34", gbp)
+        checkSpanned(gbp, "£12.34", "GBP12.34")
         assertTtsMarkers("GBP", "12", "34", gbp)
 
         // May be "¥1,234" or "JP¥1,234".
         val jpy = TransitCurrency.JPY(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(jpy, Matchers.endsWith("¥1,234"))
+        checkSpanned(jpy, "¥1,234", "JPY1,234")
         assertTtsMarkers("JPY", "1234", null, jpy)
     }
 
@@ -136,17 +135,16 @@ class TransitCurrencyTest : BaseInstrumentedTest() {
 
         // May be "$12.34", "U$12.34" or "US$12.34".
         val usd = TransitCurrency.USD(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(usd, Matchers.endsWith("$12.34"))
+        checkSpanned(usd, listOf("$12.34", "U$12.34", "US$12.34"), "USD12.34")
         assertTtsMarkers("USD", "12", "34", usd)
 
         // May be "A$12.34" or "AU$12.34".
         val aud = TransitCurrency.AUD(1234).formatCurrencyString(true).spanned
-        assertSpannedThat(aud, Matchers.startsWith("A"))
-        assertSpannedThat(aud, Matchers.endsWith("$12.34"))
+        checkSpanned(aud, listOf("A$12.34", "AU$12.34"), "AUD12.34")
         assertTtsMarkers("AUD", "12", "34", aud)
 
         val gbp = TransitCurrency(1234, "GBP").formatCurrencyString(true).spanned
-        assertSpannedEquals("£12.34", gbp)
+        checkSpanned(gbp, "£12.34", "GBP12.34")
         assertTtsMarkers("GBP", "12", "34", gbp)
 
 
@@ -165,25 +163,56 @@ class TransitCurrencyTest : BaseInstrumentedTest() {
         setLocale("fr-FR")
 
         val usd = TransitCurrency.USD(1234).formatCurrencyString(true).spanned
-        assertSpannedEquals("12,34 \$US", usd)
+        checkSpanned(usd, "12,34 \$US", "12,34 USD")
         assertTtsMarkers("USD", "12", "34", usd)
 
         val aud = TransitCurrency.AUD(1234).formatCurrencyString(true).spanned
-        assertSpannedEquals("12,34 \$AU", aud)
+        checkSpanned(aud, "12,34 \$AU", "12,34 AUD")
         assertTtsMarkers("AUD", "12", "34", aud)
 
         // Allow not qualifying the country code.
         val gbp = TransitCurrency(1234, "GBP").formatCurrencyString(true).spanned
-        assertSpannedThat(gbp, Matchers.startsWith("12,34 £"))
+        checkSpanned(gbp, listOf("12,34 £", "12,34 £GB", "12,34 £UK"), "12,34 GBP")
         assertTtsMarkers("GBP", "12", "34", gbp)
 
         // This may not have a proper symbol
         val jpy = TransitCurrency.JPY(1234).formatCurrencyString(true).spanned
-        // Accept either ordinary or narrow non-break space
-        assertSpannedThat(jpy, Matchers.anyOf(Matchers.startsWith("1 234"), Matchers.startsWith("1 234")))
+        checkSpanned(jpy, listOf("1 234 ¥", "1 234 ¥JP", "1 234 JPY"))
 
         val eur = TransitCurrency(1234, "EUR").formatCurrencyString(true).spanned
         assertSpannedEquals("12,34 €", eur)
         assertTtsMarkers("EUR", "12", "34", eur)
+    }
+
+    fun assertSpannedEquals(expected: String, actual: Spanned) {
+        // nbsp -> sp
+        val actualString = actual.toString().replace(' ', ' ').replace('\u202F', ' ')
+        assertEquals(expected, actualString)
+    }
+
+    fun assertTtsMarkers(currencyCode: String, value: String, fraction: String?, span: Spanned) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+            return
+        }
+
+        val ttsSpans = span.getSpans(0, span.length, TtsSpan::class.java)
+        assertEquals(1, ttsSpans.size)
+
+        assertEquals(TtsSpan.TYPE_MONEY, ttsSpans[0].type)
+        val bundle = ttsSpans[0].args
+        assertEquals(currencyCode, bundle.getString(TtsSpan.ARG_CURRENCY))
+        assertEquals(value, bundle.getString(TtsSpan.ARG_INTEGER_PART))
+        assertEquals(fraction, bundle.getString(TtsSpan.ARG_FRACTIONAL_PART))
+    }
+
+    fun checkSpanned(actual: Spanned, android: Collection<String>, robolectric: String? = null) {
+        // nbsp -> sp
+        val actualString = actual.toString().replace(' ', ' ').replace('\u202F', ' ')
+        val expected = if (isUnitTest) android + listOfNotNull(robolectric) else android
+        assertTrue (actual = actualString in expected, message = "Expected one of $expected, got $actualString")
+    }
+
+    fun checkSpanned(actual: Spanned, android: String, robolectric: String? = null) {
+        checkSpanned(actual, listOf(android), robolectric)
     }
 }
