+++ src/androidUnitTest/kotlin/au/id/micolous/metrodroid/test/BaseInstrumentedTestPlatform.kt
@@ -43,6 +43,8 @@ actual abstract class BaseInstrumentedTestPlatform {
     val context : Context
         get() = MetrodroidApplication.instance
 
+    val isUnitTest get() = true
+
     actual fun setLocale(languageTag: String) {
         LocaleTools.setLocale(languageTag, context.resources)
     }
